NICs: Show all network IPs (#2303)
When browsing a VM's network interfaces you would not be able to see
when a VM had multiple different IP addresses within an interface.
For example by having a secondary IP address.

Fixing this was simple as we already get the IP addresses from libvirt
and can just iterate over correctly. Instead of assuming interfaces
can only have one IPv4 and one IPv6 address we now make no distinction
on how many you can have per interface.

We also change the priority of the sources to be `agent > lease > arp`
to make sure that the IPs are gathered in the correct order. Prior to
this we only made sure `agent` was first and parsed the rest in 
whatever order they came in.

IPs are filtered to ensure that we don't get duplicate entries on the
interfaces as the different sources like `agent` or `lease` shows the
same IP address.

Fixes: https://issues.redhat.com/browse/RHEL-89015
Signed-off-by: Freya Gustavsson <freya@venefilyn.se>
Co-authored-by: Marius Vollmer <marius.vollmer@gmail.com>