consoles: Measure the card robustly
The code for measuring the card header was brittle and would build
invalid CSS selectors for VM names that contain funny characters such
as ".".  This would then crash and open the URL for the detached
console inside the iframe.

Let's do this more robustly by using getElementId and catching
exceptions.

Also, let's move the `preventDefault` call to the beginning of the
event handler so that nothing happens when the rest of the handler
crashes.