nics: Only optionally offer "network" type to user session nics
When adding or editing network interfaces, only offer the "Virtual
network" type when there are actually virtual networks defined in the
user session. Most of the time, there are none, and none can be
defined.

Also, since this leaves only the "user" type, don't show the "Type"
selector at all.