api: Move API for many domain actions into vmActions
We want domain actions like "shutdown" and "start" to have the proper
UI effects, like showing spinners and displaying errors.  The API
exported from "domain.ts" is only doing half the job and shouldn't be
used to start a VM from a random button, for example.  Instead, that
button should call the "vmStart" function exported from
"vmActions.tsx", which will do all the right things.

For this to work nicely, the state for whether a operation is in
progress has to be maintained globally, and not only in the React
component for the dropdown menu.