app: Move AppState into its own file and make it global
It is about to get a lot bigger.

Since the store is global, we make the state global too to avoid
sweeping gratuitious code changes. (We would need to pass the state
into every API function.)

Keeping the state in a React context instead of a global constant is
theoretical cleaner but has no practical advantages for us. We can
always move it back into a context when that becomes necessary.