state: Keep list of VMs in the state instead of the store
Immutable data structures are fine, and we keep them. The goal here is
to get rid of the Redux action/reducer circus. It is onerous to add or
change actions. Also, the circus is only weakly type-checked because
it seemed to be too much work at the time.

Noteworthy changes:

 - updateVm now takes separate "key" and "props" arguments. This
   allows one to simply write "updateVm(vm, { error: null })", for
   example.

 - The job of updateDomainSnapshots is now done via updateVm, which is
   perfectly suited to it.

 - undefineVm and deleteUnlistedVms now only work with ids and not
   names.

 - undefineVm has no transientOnly parameter anymore.

 - The reducer would also do business logic like clear usage data and
   compute cpu usage.  This is now done on the business side of the
   code, when detecting a shutdown machine and when processing usage
   samples.