xml: Redo all the parsing with new helper functions
This removes code noise and makes the type checking more accurate.

The main benefit is to replace code like this:

    const parentElem = getSingleOptionalElem(snapElem, 'parent');
    const parentName = parentElem?.getElementsByTagName("name")[0].childNodes[0].nodeValue;

with

    const parentName = get_text(snapElem, "parent", "name");

There were 50 or so instances like this.

Another benefit is that TypeScript assumes that

    foo.getElementsByTagName("name")[0]

is always a valid "Element". This is not true, of course, when "name"
is missing.  The new get_child, etc functions protect against this.

Also, getElementsByTagName searches recursively and will find indirect
children.  The following cases had to be corrected:

   domain/devices/cid   ->   domain/devices/vsock/cid
   volume/format/type   ->   volume/format/target/type

Still one more benefit is that get_children returns a real Array
instead of a HTMLCollection.  TypeScript doesn't let us use
HTMLCollections in "for of" loops, unfortunately, and HTMLCollections
don't have nice methods like "map".

All in all, the new functions make parsing more concise and
safer. They might be a bit slower.

This also parses the loader capabilities instead of leaving them as a
raw HTMLCollection.